+++ src/test/java/com/cluttered/cryptocurrency/ann/neuron/NeuronBuilderTest.java
@@ -6,12 +6,13 @@ import mockit.Deencapsulation;
 import mockit.Mocked;
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 
-import static com.cluttered.cryptocurrency.ann.MathConstants.ONE;
-import static com.cluttered.cryptocurrency.ann.MathConstants.ZERO;
 import static com.cluttered.cryptocurrency.ann.activation.Activation.*;
+import static java.math.BigDecimal.ONE;
+import static java.math.BigDecimal.ZERO;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
@@ -23,8 +24,8 @@ public class NeuronBuilderTest {
     public void testCreate() {
         final NeuronBuilder neuronBuilder = NeuronBuilder.create();
 
-        final BigFloat privateBias = Deencapsulation.getField(neuronBuilder, "bias");
-        final List<BigFloat> privateWeights = Deencapsulation.getField(neuronBuilder, "weights");
+        final BigDecimal privateBias = Deencapsulation.getField(neuronBuilder, "bias");
+        final List<BigDecimal> privateWeights = Deencapsulation.getField(neuronBuilder, "weights");
         final Activation privateActivation = Deencapsulation.getField(neuronBuilder, "activation");
 
         assertThat(privateBias).isEqualTo(ZERO);
@@ -33,21 +34,21 @@ public class NeuronBuilderTest {
     }
 
     @Test
-    public void testBias(@Mocked final BigFloat bias) {
+    public void testBias(@Mocked final BigDecimal bias) {
         final NeuronBuilder neuronBuilder = NeuronBuilder.create().bias(bias);
 
-        final BigFloat privateBias = Deencapsulation.getField(neuronBuilder, "bias");
+        final BigDecimal privateBias = Deencapsulation.getField(neuronBuilder, "bias");
 
         assertThat(privateBias).isEqualTo(bias);
     }
 
     @Test
-    public void testWeights(@Mocked final BigFloat weight1,
-                            @Mocked final BigFloat weight2,
-                            @Mocked final BigFloat weight3) {
+    public void testWeights(@Mocked final BigDecimal weight1,
+                            @Mocked final BigDecimal weight2,
+                            @Mocked final BigDecimal weight3) {
         final NeuronBuilder neuronBuilder = NeuronBuilder.create().weights(weight1, weight2, weight3);
 
-        final List<BigFloat> privateWeights = Deencapsulation.getField(neuronBuilder, "weights");
+        final List<BigDecimal> privateWeights = Deencapsulation.getField(neuronBuilder, "weights");
 
         assertThat(privateWeights).containsExactly(weight1, weight2, weight3);
     }
@@ -80,8 +81,8 @@ public class NeuronBuilderTest {
     }
 
     @Test
-    public void testBuild(@Mocked final BigFloat bias,
-                          @Mocked final List<BigFloat> weights,
+    public void testBuild(@Mocked final BigDecimal bias,
+                          @Mocked final List<BigDecimal> weights,
                           @Mocked final Activation activation) {
         final Neuron neuron = NeuronBuilder.create()
                 .bias(bias)
@@ -89,8 +90,8 @@ public class NeuronBuilderTest {
                 .activation(activation)
                 .build();
 
-        final BigFloat privateBias = Deencapsulation.getField(neuron, "bias");
-        final List<BigFloat> privateWeights = Deencapsulation.getField(neuron, "weights");
+        final BigDecimal privateBias = Deencapsulation.getField(neuron, "bias");
+        final List<BigDecimal> privateWeights = Deencapsulation.getField(neuron, "weights");
         final Activation privateActivation = Deencapsulation.getField(neuron, "activation");
 
         assertThat(privateBias).isEqualTo(bias);
